requirements.txt
anyio==3.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
filelock==3.12.0
h11==0.14.0
httpcore==0.17.1
httpx==0.24.1
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.3.0
numpy==1.24.3
networkx==3.1
pycparser==2.21
python-dotenv==1.0.0
python-telegram-bot==20.3
requests==2.30.0
sniffio==1.3.0
srt==3.5.3
sympy==1.12
-f https://download.pytorch.org/whl/torch_stable.html
torch==2.0.1+cpu
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==2.0.2
vosk==0.3.45
websockets==11.0.3